moving entities to new directory
