+++ core/database/src/main/java/org/mozilla/social/core/database/dao/RelationshipsDao.kt
@@ -3,7 +3,7 @@ package org.mozilla.social.core.database.dao
 import androidx.room.Dao
 import androidx.room.Query
 import kotlinx.coroutines.flow.Flow
-import org.mozilla.social.core.database.model.DatabaseRelationship
+import org.mozilla.social.core.database.model.entities.DatabaseRelationship
 
 @Dao
 interface RelationshipsDao : BaseDao<DatabaseRelationship> {
