+++ core/database/src/main/java/org/mozilla/social/core/database/model/entities/DatabaseStatus.kt
@@ -1,10 +1,17 @@
-package org.mozilla.social.core.database.model
+package org.mozilla.social.core.database.model.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.datetime.Instant
+import org.mozilla.social.core.database.model.DatabaseApplication
+import org.mozilla.social.core.database.model.DatabaseAttachment
+import org.mozilla.social.core.database.model.DatabaseCard
+import org.mozilla.social.core.database.model.DatabaseEmoji
+import org.mozilla.social.core.database.model.DatabaseHashTag
+import org.mozilla.social.core.database.model.DatabaseMention
+import org.mozilla.social.core.database.model.DatabaseStatusVisibility
 
 /**
  * Represents a status posted by an account.
