+++ core/database/src/main/java/org/mozilla/social/core/database/model/entities/accountCollections/Follower.kt
@@ -1,12 +1,12 @@
-package org.mozilla.social.core.database.model.accountCollections
+package org.mozilla.social.core.database.model.entities.accountCollections
 
 import androidx.room.ColumnInfo
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Relation
-import org.mozilla.social.core.database.model.DatabaseAccount
-import org.mozilla.social.core.database.model.DatabaseRelationship
+import org.mozilla.social.core.database.model.entities.DatabaseAccount
+import org.mozilla.social.core.database.model.entities.DatabaseRelationship
 
 /**
  * @param accountId the account ID of the user we are examining
