+++ core/database/src/main/java/org/mozilla/social/core/database/model/entities/statusCollections/HashTagTimelineStatus.kt
@@ -1,17 +1,20 @@
-package org.mozilla.social.core.database.model.statusCollections
+package org.mozilla.social.core.database.model.entities.statusCollections
 
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.ForeignKey
-import androidx.room.PrimaryKey
 import androidx.room.Relation
-import org.mozilla.social.core.database.model.DatabaseAccount
-import org.mozilla.social.core.database.model.DatabasePoll
-import org.mozilla.social.core.database.model.DatabaseStatus
+import org.mozilla.social.core.database.model.entities.DatabaseAccount
+import org.mozilla.social.core.database.model.entities.DatabasePoll
+import org.mozilla.social.core.database.model.entities.DatabaseStatus
 import org.mozilla.social.core.database.model.wrappers.StatusWrapper
 
 @Entity(
-    tableName = "favoritesTimeline",
+    tableName = "hashTagTimeline",
+    primaryKeys = [
+        "statusId",
+        "hashTag",
+    ],
     foreignKeys = [
         ForeignKey(
             entity = DatabaseStatus::class,
@@ -57,10 +60,9 @@ import org.mozilla.social.core.database.model.wrappers.StatusWrapper
         ),
     ]
 )
-data class FavoritesTimelineStatus(
-    @PrimaryKey
+data class HashTagTimelineStatus(
     val statusId: String,
-    val position: Int,
+    val hashTag: String,
     val accountId: String,
     val pollId: String?,
     val boostedStatusId: String?,
@@ -68,9 +70,9 @@ data class FavoritesTimelineStatus(
     val boostedPollId: String?,
 )
 
-data class FavoritesTimelineStatusWrapper(
+data class HashTagTimelineStatusWrapper(
     @Embedded
-    val favoritesTimelineStatus: FavoritesTimelineStatus,
+    val hashTagTimelineStatus: HashTagTimelineStatus,
     @Relation(
         parentColumn = "statusId",
         entityColumn = "statusId",
@@ -103,7 +105,7 @@ data class FavoritesTimelineStatusWrapper(
     val boostedPoll: DatabasePoll?,
 )
 
-fun FavoritesTimelineStatusWrapper.toStatusWrapper(): StatusWrapper =
+fun HashTagTimelineStatusWrapper.toStatusWrapper(): StatusWrapper =
     StatusWrapper(
         status = status,
         account = account,
