+++ core/database/src/main/java/org/mozilla/social/core/database/model/wrappers/StatusWrapper.kt
@@ -2,9 +2,9 @@ package org.mozilla.social.core.database.model.wrappers
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import org.mozilla.social.core.database.model.DatabaseAccount
-import org.mozilla.social.core.database.model.DatabasePoll
-import org.mozilla.social.core.database.model.DatabaseStatus
+import org.mozilla.social.core.database.model.entities.DatabaseAccount
+import org.mozilla.social.core.database.model.entities.DatabasePoll
+import org.mozilla.social.core.database.model.entities.DatabaseStatus
 
 data class StatusWrapper(
     @Embedded
