+++ core/repository/mastodon/src/main/java/org/mozilla/social/core/repository/mastodon/TimelineRepository.kt
@@ -17,12 +17,12 @@ import org.mozilla.social.core.database.dao.FederatedTimelineStatusDao
 import org.mozilla.social.core.database.dao.HashTagTimelineStatusDao
 import org.mozilla.social.core.database.dao.HomeTimelineStatusDao
 import org.mozilla.social.core.database.dao.LocalTimelineStatusDao
-import org.mozilla.social.core.database.model.statusCollections.AccountTimelineStatusWrapper
-import org.mozilla.social.core.database.model.statusCollections.FederatedTimelineStatusWrapper
-import org.mozilla.social.core.database.model.statusCollections.HashTagTimelineStatusWrapper
-import org.mozilla.social.core.database.model.statusCollections.HomeTimelineStatusWrapper
-import org.mozilla.social.core.database.model.statusCollections.LocalTimelineStatusWrapper
-import org.mozilla.social.core.database.model.statusCollections.toStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.AccountTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.FederatedTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.HashTagTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.HomeTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.LocalTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.toStatusWrapper
 import org.mozilla.social.core.model.Status
 import org.mozilla.social.core.model.StatusVisibility
 import org.mozilla.social.core.model.AccountTimelineType
