+++ core/repository/mastodon/src/main/java/org/mozilla/social/core/repository/mastodon/model/account/DatabaseToExternal.kt
@@ -1,6 +1,6 @@
 package org.mozilla.social.core.repository.mastodon.model.account
 
-import org.mozilla.social.core.database.model.DatabaseRelationship
+import org.mozilla.social.core.database.model.entities.DatabaseRelationship
 import org.mozilla.social.core.model.Relationship
 
 fun DatabaseRelationship.toExternal(): Relationship =
