+++ core/usecase/mastodon/src/main/java/org/mozilla/social/core/usecase/mastodon/timeline/RefreshFederatedTimeline.kt
@@ -6,7 +6,7 @@ import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import kotlinx.coroutines.delay
 import org.mozilla.social.common.Rel
-import org.mozilla.social.core.database.model.statusCollections.FederatedTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.FederatedTimelineStatusWrapper
 import org.mozilla.social.core.repository.mastodon.AccountRepository
 import org.mozilla.social.core.repository.mastodon.DatabaseDelegate
 import org.mozilla.social.core.repository.mastodon.TimelineRepository
