+++ core/usecase/mastodon/src/main/java/org/mozilla/social/core/usecase/mastodon/timeline/RefreshLocalTimeline.kt
@@ -6,7 +6,7 @@ import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import kotlinx.coroutines.delay
 import org.mozilla.social.common.Rel
-import org.mozilla.social.core.database.model.statusCollections.LocalTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.LocalTimelineStatusWrapper
 import org.mozilla.social.core.repository.mastodon.AccountRepository
 import org.mozilla.social.core.repository.mastodon.DatabaseDelegate
 import org.mozilla.social.core.repository.mastodon.TimelineRepository
