+++ feature/feed/src/main/java/org/mozilla/social/feed/remoteMediators/FederatedTimelineRemoteMediator.kt
@@ -4,7 +4,7 @@ import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
-import org.mozilla.social.core.database.model.statusCollections.FederatedTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.FederatedTimelineStatusWrapper
 import org.mozilla.social.core.usecase.mastodon.timeline.RefreshFederatedTimeline
 
 @OptIn(ExperimentalPagingApi::class)
