+++ feature/feed/src/main/java/org/mozilla/social/feed/remoteMediators/HomeTimelineRemoteMediator.kt
@@ -4,7 +4,7 @@ import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
-import org.mozilla.social.core.database.model.statusCollections.HomeTimelineStatusWrapper
+import org.mozilla.social.core.database.model.entities.statusCollections.HomeTimelineStatusWrapper
 import org.mozilla.social.core.usecase.mastodon.timeline.RefreshHomeTimeline
 
 @OptIn(ExperimentalPagingApi::class)
